refactor: extract submit modal
